pyproject.toml
[project]
name = "faster-watching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.182.0",
    "google-generativeai>=0.8.5",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "youtube-transcript-api>=1.2.2",
]
